Misc cleanup
Refactored the filter/download logic so regardless of GUI or CLI it takes place in the same functions.

Fixed the bug where when I set globalWindow to None but it was a local variable and not the global instance.

When you start the program it opens a command line window. If you are running with a GUI I hide that window and restore it after quiting.